Add files via upload
The main purpose of this mobile application is to meet the needs of companies in the companies in the quality control operation. By automating the quality control at reception, we will both eliminate the possible errors in manual picking, reduce costs and optimise the process.

By scanning a simple QR code, we will be able to inspect the goods quality at receipt. This simple and fast mobile application, will facilitate the inspection procedures by sharing of reliable information within the system..
Finally, seeing the increasing competition in the market, organisations are striving to be more innovative in their products and services, therefore investing in such a solution will be purely beneficial for any company.